+++ market/migrations/0012_tradeitem_group_id_tradeitem_market_group_id.py
@@ -13,13 +13,13 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='tradeitem',
             name='group_id',
-            field=models.IntegerField(db_index=True, default=0),
+            field=models.BigIntegerField(db_index=True, default=0),
             preserve_default=False,
         ),
         migrations.AddField(
             model_name='tradeitem',
             name='market_group_id',
-            field=models.IntegerField(db_index=True, default=0),
+            field=models.BigIntegerField(db_index=True, default=0),
             preserve_default=False,
         ),
     ]
