+++ sde/migrations/0009_alter_npccorporation_faction_id.py
@@ -13,6 +13,6 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='npccorporation',
             name='faction_id',
-            field=models.IntegerField(blank=True, default=None, null=True),
+            field=models.BigIntegerField(blank=True, default=None, null=True),
         ),
     ]
